tabs created based on resources and current tab is defined
tabs are now created based on resources and will change when new objects are added.
if a tab has the class active tab it will get a underline styling. Added a variable that keeps track of the current tab and a function that toggles the activeTab class on the button when it is the current tab.